Split continues — caption tab & caption model moved to core/ .
Continues the split. CaptionTab (1,694 lines) → core/tabs/caption_tab.py and
CaptionModel (282 lines) → core/models/caption_model.py. Pure move: signatures
are preserved; the main file imports them. Monolith: 15,299 → 13,348 lines.
Per-module refactoring is deferred.
